% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFSDataExtract.R
\name{combineMSFSDailyFiles}
\alias{combineMSFSDailyFiles}
\title{Combine Daily MSFS Files}
\usage{
combineMSFSDailyFiles(
  InpDir = paste0(Sys.getenv("USERPROFILE"),
    "\\\\Callodine Capital Management, LP\\\\TWood - Documents\\\\MSFS_HoldingsData_Daily\\\\")
)
}
\description{
Combines csv files containing daily MSFS Data
}
