% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFSDataExtract.R
\name{unpackMSFSFile}
\alias{unpackMSFSFile}
\title{Unpack MSFS "CCMF_DAILY_GL_SEC" Report from Portal into Daily Data File}
\usage{
unpackMSFSFile(
  InpDir,
  OutDir = paste0(Sys.getenv("USERPROFILE"),
    "\\\\Callodine Capital Management, LP\\\\TWood - Documents\\\\MSFS_HoldingsData_Daily\\\\")
)
}
\description{
Reads Latest FTP File from MSFS for Callodine Dashboard
}
