% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFSDataExtract.R
\name{unpackMSFSFTP}
\alias{unpackMSFSFTP}
\title{Unpack MSFS "CCMF_DAILY_GL_SEC" Report from FTP into Daily Data File}
\usage{
unpackMSFSFTP(
  FTPDir = paste0(Sys.getenv("USERPROFILE"),
    "\\\\Callodine Capital Management, LP\\\\TWood - Documents\\\\MSFS_FTP\\\\"),
  OutDir = paste0(Sys.getenv("USERPROFILE"),
    "\\\\Callodine Capital Management, LP\\\\TWood - Documents\\\\MSFS_HoldingsData_Daily\\\\")
)
}
\description{
Reads Latest FTP File from MSFS for Callodine Dashboard
}
